chore(table-of-contents): rename utils file
