+++ src/components/TableOfContents/TableOfContentsMobile.tsx
@@ -11,7 +11,7 @@ import React from "react"
 import { MdExpandMore } from "react-icons/md"
 import Translation from "../Translation"
 import ItemsList from "./ItemsList"
-import { Item, outerListProps } from "./tableOfContents.utils"
+import { Item, outerListProps } from "./utils"
 
 export interface IPropsTableOfContentsMobile {
   items?: Array<Item>
